Se cambia la ubicacion de los componentes de AdminPanel a su lugar correcto y se finaliza la restructuracion de front al 100%
