+++ FRONT/src/Components/index.js
@@ -10,6 +10,3 @@ export * from './SearchBar'
 
 export { Login, Register } from './Auth'
 export * from './Detail/Detail'
-
-
-
