+++ FRONT/src/Routes/AppRouter.jsx
@@ -4,8 +4,8 @@ import { PublicRoute } from "./PublicRoute"
 import { PrivateRoute } from "./PrivateRoute"
 import { Home } from "../Layouts"
 import { Almuerzo, Cena, Desayuno, Detail, Login, Mediatarde, Register } from "../Components"
-import AdminPanel from "./AdminPanel/AdminPanel"
-import CrearReceta from "./AdminPanel/gestionRecetas/CrearReceta"
+import AdminPanel from "../Layouts/AdminPanel"
+import CrearReceta from "../Components/AdminPanel/gestionRecetas/CrearReceta"
 import MyAccount from "../Components/MyAccount"
 import Favs from "../Layouts/Favs"
